Only do roundtrip for ODF files.
svn path=/trunk/koffice/; revision=1172673

+++ tools/scripts/profileOfficeFileLoading.py
@@ -148,9 +148,14 @@ def profile(dir, file, logger):
 
 	# profile
 	(fileno, tmpfilename) = tempfile.mkstemp()
-	(roundtripfd, roundtripfilename) = tempfile.mkstemp()
-	args = ["--benchmark-loading", "--profile-filename", tmpfilename,
-		"--roundtrip-filename", roundtripfilename,
+	roundtripfilename = None
+	args = []
+	# in case of ODF file, do a roundtrip
+        m = re.match('.*(\.od.)$', file)
+	if m:
+		(roundtripfd, roundtripfilename) = tempfile.mkstemp(m.group(1))
+		args += ["--roundtrip-filename", roundtripfilename]
+	args += ["--benchmark-loading", "--profile-filename", tmpfilename,
 		"--nocrashhandler", file]
 	r = runCommand(exepath, args, False)
 	outfile = os.fdopen(fileno, 'r')
@@ -169,7 +174,8 @@ def profile(dir, file, logger):
 		logger.failTest(r.backtrace)
 
 	os.remove(tmpfilename)
-	os.remove(roundtripfilename)
+	if roundtripfilename:
+		os.remove(roundtripfilename)
 
 	logger.endTest(int((r.utime + r.stime)*1000))
 	return r
